requirements.txt
asgiref==3.8.1
Django==5.1.5
django-extensions==3.2.3
djangorestframework==3.15.2
# Using the latest version from your list and removing duplicate
djangorestframework-simplejwt==5.4.0
fuzzywuzzy==0.18.0
graphviz==0.20.3
inflection==0.5.1
# Using python-Levenshtein as it pairs with fuzzywuzzy and has binary support
python-Levenshtein==0.26.1
# Using binary version as requested
psycopg2-binary==2.9.10
PyJWT==2.10.1
PyYAML==6.0.2
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
gunicorn==21.2.0
django-cors-headers==4.3.1
django-filter==23.5